Edited app flow based on results of App Store review.

+++ RecipeGroups/Views and ViewControllers/My Groups/GroupDetailTableViewController.swift
@@ -7,8 +7,9 @@
 //
 
 import UIKit
+import MessageUI
 
-class GroupDetailTableViewController: UITableViewController {
+class GroupDetailTableViewController: UITableViewController, MFMessageComposeViewControllerDelegate {
     
     var group: Group?
 
@@ -19,19 +20,35 @@ class GroupDetailTableViewController: UITableViewController {
     }
 
     override func numberOfSections(in tableView: UITableView) -> Int {
-        return 3
+        guard let group = group else { return 0 }
+        if group.password != nil {
+            return 4
+        } else {
+            return 3
+        }
     }
 
     override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         guard let group = group else { return 0 }
         
-        switch section {
-        case 0, 1:
-            return 1
-        case 2:
-            return group.members.count
-        default:
-            return 0
+        if group.password != nil {
+            switch section {
+            case 0, 1, 2:
+                return 1
+            case 3:
+                return group.members.count
+            default:
+                return 0
+            }
+        } else {
+            switch section {
+            case 0, 1:
+                return 1
+            case 2:
+                return group.members.count
+            default:
+                return 0
+            }
         }
     }
 
@@ -44,9 +61,14 @@ class GroupDetailTableViewController: UITableViewController {
                 cell.groupLabel.text = group.name
                 return cell
             }
-        } else if indexPath.section == 1 || indexPath.section == 2 {
+        } else if indexPath.section == 1 && group.password != nil {
+            if let cell = tableView.dequeueReusableCell(withIdentifier: "GroupPasswordCell", for: indexPath) as? GroupPasswordTableViewCell {
+                cell.passwordLabel.text = group.password
+                return cell
+            }
+        } else {
             if let cell = tableView.dequeueReusableCell(withIdentifier: "GroupMemberCell", for: indexPath) as? GroupMemberTableViewCell {
-                if indexPath.section == 1 {
+                if (indexPath.section == 1 && group.password == nil) || (indexPath.section == 2 && group.password != nil) {
                     cell.memberLabel.text = group.creator
                 } else {
                     cell.memberLabel.text = group.members[indexPath.row]
@@ -63,44 +85,81 @@ class GroupDetailTableViewController: UITableViewController {
         let groupName = "Group Name"
         let creator = "Creator"
         let groupMembers = "Group Members"
+        let groupPassword = "Group Password"
         
         switch section {
         case 0:
             return groupName
         case 1:
+            if group?.password != nil {
+                return groupPassword
+            }
             return creator
+        case 2:
+            if group?.password != nil {
+                return creator
+            }
+            return groupMembers
         default:
             return groupMembers
         }
     }
     
     func setupView() {
-        if group?.creator == UserController.shared.user?.email {
-            navigationItem.rightBarButtonItem = UIBarButtonItem(title: "Delete Group", style: .plain, target: self, action: #selector(deleteGroupTapped))
+        if group?.password != nil {
+            navigationItem.rightBarButtonItem = UIBarButtonItem(barButtonSystemItem: .action, target: self, action: #selector(creatorGroupActionTapped))
         } else {
             navigationItem.rightBarButtonItem = UIBarButtonItem(title: "Leave Group", style: .plain, target: self, action: #selector(leaveGroupTapped))
         }
     }
     
-    @objc func deleteGroupTapped() {
-        let alertController = UIAlertController(title: "Delete Group", message: "Are you sure you want to delete this group?", preferredStyle: .alert)
-        let cancelAction = UIAlertAction(title: "Cancel", style: .cancel, handler: nil)
-        
-        // Send a request to the server to delete the group if "Delete Group" button is tapped
-        let deleteGroupAction = UIAlertAction(title: "Delete Group", style: .default) { (_) in
-            GroupController.shared.deleteGroup(named: self.group!.name, completion: { (result) in
-                if result == "Success" {
-                    GroupController.shared.processDeletedGroup(named: self.group!.name)
-                    self.navigationController?.popViewController(animated: true)
-                }
-            })
-        }
-        
-        alertController.addAction(cancelAction)
-        alertController.addAction(deleteGroupAction)
-        present(alertController, animated: true, completion: nil)
+    func messageComposeViewController(_ controller: MFMessageComposeViewController, didFinishWith result: MessageComposeResult) {
+        self.dismiss(animated: true, completion: nil)
     }
     
+    @objc func creatorGroupActionTapped() {
+        guard let groupName = group?.name else { return }
+        guard let password = group?.password else { return }
+        
+        let alertController = UIAlertController(title: nil, message: nil, preferredStyle: .actionSheet)
+        let inviteToJoinAction = UIAlertAction(title: "Invite Someone to Join", style: .default) { (_) in
+            let composeVC = MFMessageComposeViewController()
+            composeVC.messageComposeDelegate = self
+            composeVC.body = """
+                Join my Recipe Group!
+                Group name: \(groupName)
+                Password: \(password)
+                """
+            if MFMessageComposeViewController.canSendText() {
+                self.present(composeVC, animated: true, completion: nil)
+            }
+        }
+        let deleteGroupAction = UIAlertAction(title: "Delete Group", style: .default) { (_) in
+            
+            // Create a second UIAlertController to confirm that the user wants to delete the group
+            let alertController = UIAlertController(title: "Delete Group", message: "Are you sure you want to delete this group?", preferredStyle: .alert)
+            let cancelAction = UIAlertAction(title: "Cancel", style: .cancel, handler: nil)
+            
+            // Send a request to the server to delete the group if "Delete Group" button is tapped
+            let deleteGroupAction = UIAlertAction(title: "Delete Group", style: .default) { (_) in
+                GroupController.shared.deleteGroup(named: self.group!.name, completion: { (result) in
+                    if result == "Success" {
+                        GroupController.shared.processDeletedGroup(named: self.group!.name)
+                        self.navigationController?.popViewController(animated: true)
+                    }
+                })
+            }
+            alertController.addAction(cancelAction)
+            alertController.addAction(deleteGroupAction)
+            self.present(alertController, animated: true, completion: nil)
+        }
+        
+        let cancelAction = UIAlertAction(title: "Cancel", style: .cancel, handler: nil)
+        alertController.addAction(inviteToJoinAction)
+        alertController.addAction(deleteGroupAction)
+        alertController.addAction(cancelAction)
+        self.present(alertController, animated: true, completion: nil)
+    }
     
     @objc func leaveGroupTapped() {
         let alertController = UIAlertController(title: "Leave Group", message: "Are you sure you want to leave this group?", preferredStyle: .alert)
@@ -120,51 +179,5 @@ class GroupDetailTableViewController: UITableViewController {
         alertController.addAction(leaveGroupAction)
         present(alertController, animated: true, completion: nil)
     }
-    
-
-    /*
-    // Override to support conditional editing of the table view.
-    override func tableView(_ tableView: UITableView, canEditRowAt indexPath: IndexPath) -> Bool {
-        // Return false if you do not want the specified item to be editable.
-        return true
-    }
-    */
-
-    /*
-    // Override to support editing the table view.
-    override func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
-        if editingStyle == .delete {
-            // Delete the row from the data source
-            tableView.deleteRows(at: [indexPath], with: .fade)
-        } else if editingStyle == .insert {
-            // Create a new instance of the appropriate class, insert it into the array, and add a new row to the table view
-        }    
-    }
-    */
-
-    /*
-    // Override to support rearranging the table view.
-    override func tableView(_ tableView: UITableView, moveRowAt fromIndexPath: IndexPath, to: IndexPath) {
-
-    }
-    */
-
-    /*
-    // Override to support conditional rearranging of the table view.
-    override func tableView(_ tableView: UITableView, canMoveRowAt indexPath: IndexPath) -> Bool {
-        // Return false if you do not want the item to be re-orderable.
-        return true
-    }
-    */
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
 
 }
